# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with filters
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: paymentType
          in: query
          schema:
            $ref: "#/components/schemas/PaymentType"
        - name: responsiblePerson
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    PaymentType:
      type: string
      enum: [cash, card, transfer]
    TransactionType:
      type: string
      enum: [income, expense]
    UserRole:
      type: string
      enum: [admin, operator]
    User:
      type: object
      required: [id, clerkUserId, name, email, role, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        name: { type: string }
        email: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt: { type: string, format: date-time }
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
...
    Transaction:
      type: object
      required:
        - id
        - clientId
        - clientName
        - type
        - amount
        - date
        - paymentType
        - responsiblePerson
        - comment
        - createdAt
      properties:
        id: { type: integer }
        clientId: { type: integer }
        clientName: { type: string }
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
          minimum: 0
        date: { type: string, format: date }
        paymentType:
          $ref: "#/components/schemas/PaymentType"
        responsiblePerson: { type: string }
        comment: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    TransactionInput:
      type: object
      required: [clientId, type, amount, date, paymentType, responsiblePerson]
      properties:
        clientId: { type: integer }
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
          exclusiveMinimum: 0
        date: { type: string, format: date }
        paymentType:
          $ref: "#/components/schemas/PaymentType"
        responsiblePerson: { type: string, minLength: 1 }
        comment: { type: string }
    TransactionUpdate:
      type: object
      properties:
        clientId: { type: integer }
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
          exclusiveMinimum: 0
        date: { type: string, format: date }
        paymentType:
          $ref: "#/components/schemas/PaymentType"
        responsiblePerson: { type: string, minLength: 1 }
        comment: { type: string }
...