# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentType } from './paymentType';
import type { TransactionType } from './transactionType';

export interface TransactionUpdate {
  clientId?: number;
  type?: TransactionType;
  /** @exclusiveMinimum 0 */
  amount?: number;
  date?: Date;
  paymentType?: PaymentType;
  /** @minLength 1 */
  responsiblePerson?: string;
  comment?: string;
}