# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  clerkUserId: string;
  name: string;
  email: string;
  role: UserRole;
  createdAt: Date;
}