# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  operator: 'operator',
} as const;