# lib/api-zod/src/generated/types/userRoleUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface UserRoleUpdate {
  role: UserRole;
}